[06-06-24]
Within the "CalcView" function, create a read-only variable named displayText and assign the mutableStateOf value of "0".

Within the "CalcView" function, after the read-only variable create a Column and give it a light grey background. Within the Column, create two Rows. Within the first row call CalcDisplay.

Within the second Row of "CalcView" create two Columns. In the second Column use the "CalcOperationButton" to create the four calculation operators.

Within "CalcView" in the first of the Columns created in the last step create a for loop that will count from 7 to 1 in steps of 3. Within the loop call the CalcRow method with the current loop value, 3 buttons, and the display text.