16-06-24
Inside of the If state just created: 1. Create a mutable variable named answer and assign a value of 0. 2. Create a when statement and use the operator variable. 3. Inside the when statement, use the strings "+', "-", "*", and "/" to assign the outcome of the operation to the answer variable. 4. Assign the answer variable to the displayText variable.