16-06-24
Inside the number press function right under the if statement do the following: 1. If operation variable is not blank and the complete variable is not true, multiply the right number by 10 and add the btnNum variable. Store the result into the rightNumber variable. 2. If operation variable is blank and the complete variable is not true, multiply the left number by 10 and add the btnNum variable. Store the result into the leftNumber variable.