16-06-24
In the CalcRow and the CalcNumericButton functions, change the display parameter name to onPress and the type to the following: (number: Int) -> Unit
In the CalcOperationButton function, change the display parameter name to onPress and the type to the following: onPress: (operation: String) -> Unit
In the CalcEqualsButton function, change the display parameter to onpress and the type to the following: () -> Unit
Back in the CalcView function, where CalcRow is called, change the display variable to a lambda that takes an integer named number and executes the numberPress function, passing in the number parameter. This lambda is assigned to the onPress event of CalcRow.
In the CalcView function, where CalcNumericButton is called, put in the same lambda for the onPress event.
In the CalcView function, where the CalcEqualButton is called, pass a lambda that simply calls equalsPress for the onPress event.
In the CalcView function, where we call the CalcOperation function. Create a lambda that takes in a string parameter named op and than calls the operationPress function passing the op parameter.
In the CalcRow function, remove the displayTest variable and pass the onPress variable instead.